+++ _swaptions/src/HJM.h
@@ -7,24 +7,24 @@
 #include <sys/time.h>
 
 
-FTYPE RanUnif( long *s );
-void RanUnif_vector( long *s , int iFactors , int iN ,int  BLOCKSIZE, FTYPE **randZ);
+FTYPE RanUnif( int *s );
+void RanUnif_vector( int *s , int iFactors , int iN ,int  BLOCKSIZE, FTYPE **randZ);
 
 FTYPE CumNormalInv( FTYPE u );
-void CumNormalInv_vector( FTYPE* u ,FTYPE* output ,unsigned long int gvl);
+void CumNormalInv_vector( FTYPE* u ,FTYPE* output ,unsigned int gvl);
 
 void icdf_SSE(const int N, FTYPE *in, FTYPE *out);
 void icdf_baseline(const int N, FTYPE *in, FTYPE *out);
 int HJM_SimPath_Forward_SSE(FTYPE **ppdHJMPath, int iN, int iFactors, FTYPE dYears, FTYPE *pdForward, FTYPE *pdTotalDrift,
-			    FTYPE **ppdFactors, long *lRndSeed);
+			    FTYPE **ppdFactors, int *lRndSeed);
 int Discount_Factors_SSE(FTYPE *pdDiscountFactors, int iN, FTYPE dYears, FTYPE *pdRatePath);
 int Discount_Factors_opt(FTYPE *pdDiscountFactors, int iN, FTYPE dYears, FTYPE *pdRatePath);
 
 
 int HJM_SimPath_Forward_Blocking_SSE(FTYPE **ppdHJMPath, int iN, int iFactors, FTYPE dYears, FTYPE *pdForward, FTYPE *pdTotalDrift,
-			    FTYPE **ppdFactors, long *lRndSeed, int BLOCKSIZE);
+			    FTYPE **ppdFactors, int *lRndSeed, int BLOCKSIZE);
 int HJM_SimPath_Forward_Blocking(FTYPE **ppdHJMPath, int iN, int iFactors, FTYPE dYears, FTYPE *pdForward, FTYPE *pdTotalDrift,
-			    FTYPE **ppdFactors, long *lRndSeed, int BLOCKSIZE);
+			    FTYPE **ppdFactors, int *lRndSeed, int BLOCKSIZE);
 
 
 int Discount_Factors_Blocking(FTYPE *pdDiscountFactors, int iN, FTYPE dYears, FTYPE *pdRatePath, int BLOCKSIZE);
@@ -50,8 +50,8 @@ int HJM_Swaption_Blocking_SSE(FTYPE *pdSwaptionPrice, //Output vector that will 
 			      FTYPE *pdYield, 
 			      FTYPE **ppdFactors,
 			      //Simulation Parameters
-			      long iRndSeed, 
-			      long lTrials, int blocksize, int tid);
+			      int iRndSeed, 
+			      int lTrials, int blocksize, int tid);
  
 int HJM_Swaption_Blocking(FTYPE *pdSwaptionPrice, //Output vector that will store simulation results in the form:
 			                              //Swaption Price
@@ -71,12 +71,12 @@ int HJM_Swaption_Blocking(FTYPE *pdSwaptionPrice, //Output vector that will stor
 			      FTYPE *pdYield, 
 			      FTYPE **ppdFactors,
 			      //Simulation Parameters
-			      long *iRndSeed, 
-			      long lTrials, int blocksize, int tid);
+			      int *iRndSeed, 
+			      int lTrials, int blocksize, int tid);
 
 /*
-extern "C" FTYPE *dvector( long nl, long nh );
-extern "C" FTYPE **dmatrix( long nrl, long nrh, long ncl, long nch );
-extern "C" void free_dvector( FTYPE *v, long nl, long nh );
-extern "C" void free_dmatrix( FTYPE **m, long nrl, long nrh, long ncl, long nch );
+extern "C" FTYPE *dvector( int nl, int nh );
+extern "C" FTYPE **dmatrix( int nrl, int nrh, int ncl, int nch );
+extern "C" void free_dvector( FTYPE *v, int nl, int nh );
+extern "C" void free_dmatrix( FTYPE **m, int nrl, int nrh, int ncl, int nch );
 */
